initial split of Note types into saved object and api types
Also removed unused types
