+++ x-pack/plugins/security_solution/public/timelines/components/open_timeline/types.ts
@@ -8,7 +8,7 @@
 import type React from 'react';
 import type { AllTimelinesVariables } from '../../containers/all';
 import type { TimelineModel } from '../../store/timeline/model';
-import type { NoteResult } from '../../../../common/types/timeline/note';
+import type { NoteResult } from '../../../../common/types/timeline/note/api';
 import type {
   RowRendererId,
   SingleTimelineResolveResponse,
