+++ x-pack/plugins/security_solution/server/lib/timeline/routes/timelines/create_timelines/helpers.ts
@@ -12,7 +12,7 @@ import { timeline as timelineLib, pinnedEvent as pinnedEventLib } from '../../..
 import type { FrameworkRequest } from '../../../../framework';
 import type { ResponseTimeline, SavedTimeline } from '../../../../../../common/types/timeline/api';
 import { persistNotes } from '../../../saved_object/notes/persist_notes';
-import type { NoteResult } from '../../../../../../common/types/timeline/note';
+import type { NoteResult } from '../../../../../../common/types/timeline/note/api';
 
 interface CreateTimelineProps {
   frameworkRequest: FrameworkRequest;
