+++ x-pack/plugins/security_solution/server/lib/timeline/schemas/notes/index.ts
@@ -7,7 +7,7 @@
 
 import * as runtimeTypes from 'io-ts';
 import { unionWithNullType } from '../../../../../common/utility_types';
-import { NoteServerRepresentation } from '../../../../../common/types/timeline/note';
+import { NoteServerRepresentation } from '../../../../../common/types/timeline/note/api';
 
 export const eventNotes = unionWithNullType(runtimeTypes.array(NoteServerRepresentation));
 export const globalNotes = unionWithNullType(runtimeTypes.array(NoteServerRepresentation));
